folder/requirements.txt
yfinance==0.1.70
fredapi
google-auth-oauthlib
pandas-gbq
pyotp
robin_stocks
alpaca-trade-api
alpaca.py
talipp
fpdf
bokeh
networkx
ta
statsmodels
arch
yfinance
hurst
hurst_exponent
lightweight-charts
pywebview
PyQt5
qtpy
mplfinance
nest_asyncio

